+++ rawbuilder/dataset/DataSet.py
@@ -1,6 +1,7 @@
 import pkg_resources
 import json
 import pandas as pd
+import os
 from ..factory.Factory import Factory
 
 
@@ -74,17 +75,19 @@ class DataSet:
         except ValueError:
             raise ValueError('Schema JSON is invalid')
 
-    def build(self, export_csv=True, return_df=False):
+    def build(self, output_path: str = None, export_csv=True, return_df=False):
         """
         Build the dataset
 
         Args:
+            output_path(str): Define a path to write the files to
             export_csv(bool): Return and write the dataset as a CSV file
             return_df(bool): Return the pandas dataframe after building it.
 
         Raises:
               ValueError: If the required task is not in the schema file
               ValueError: If the return_csv or return_df are not bool
+              NotADirectoryError: If the output_path is not a valid os directory
 
         Returns:
               Pandas DataFrame (optional)
@@ -92,6 +95,9 @@ class DataSet:
 
         if not isinstance(export_csv, bool) or not isinstance(return_df, bool):
             raise ValueError('Arguments return_csv, return_df must of type bool')
+
+        if output_path is not None and not os.path.exists(output_path):
+            raise NotADirectoryError('Output path must be a directory')
 
         self._read_schema_file()
 
@@ -106,6 +112,8 @@ class DataSet:
             df[column_name] = pd.Series(data=factory.build_column(column_data_type))
 
         output_file_name = '{}_{}.csv'.format(self._config.get('task'), self._config.get('size'))
+        if output_path:
+            output_file_name = '{}/{}'.format(output_path, output_file_name)
 
         if export_csv:
             df.to_csv(output_file_name, chunksize=1000, index=False)
@@ -116,4 +124,3 @@ class DataSet:
         if return_df:
             return df
         del df
-